Fix multi-fetch
Some previous changes made multi-fetch incorrect.

Let's fix it.

+++ src/Searchlight/SqlExecutor.cs
@@ -55,14 +55,24 @@ namespace Searchlight
             // If the user wants multi-fetch to retrieve row count
             if (useMultiFetch)
             {
-                var commandClauses = sql.ResultSetClauses.Count > 0
-                    ? String.Join("\n", sql.ResultSetClauses) + "\n"
-                    : "";
-                sql.CommandText = $"SELECT * INTO #temp FROM {query.Source.TableName}{where};\n" +
-                                  $"SELECT COUNT(1) AS TotalRecords FROM #temp;\n" +
-                                  $"SELECT * FROM #temp{order}{offset};\n" + 
-                                  commandClauses +
-                                  $"DROP TABLE #temp;\n";
+                // If we're doing multi-fetch, we have to retrieve sorted/paginated records into a temp table before
+                // joining with any child collections
+                if (sql.ResultSetClauses.Count > 0)
+                {
+                    var commandClauses = sql.ResultSetClauses.Count > 0
+                        ? String.Join("\n", sql.ResultSetClauses) + "\n"
+                        : "";
+                    sql.CommandText = $"SELECT COUNT(1) AS TotalRecords FROM {query.Source.TableName}{where};\n" +
+                                      $"SELECT * INTO #temp FROM {query.Source.TableName}{where}{order}{offset};\n" +
+                                      $"SELECT * FROM #temp{order};\n" +
+                                      commandClauses +
+                                      $"DROP TABLE #temp;\n";
+                }
+                else
+                {
+                    sql.CommandText = $"SELECT COUNT(1) AS TotalRecords FROM {query.Source.TableName}{where};\n" +
+                                      $"SELECT * FROM {query.Source.TableName}{where}{order}{offset};\n";
+                }
             }
             else
             {
